Agregadas pruebas basicas de hacha de madera, piedra y pico de madera

+++ src/fiuba/algo3/tp2/PicoFino.java
@@ -6,7 +6,7 @@ public class PicoFino extends Pico {
 		super(material);
 	}
 	
-	public int getDurabilidad() {
+	public float getDurabilidad() {
 		return 1000;
 	}
